QQuickFBO: Keep devicePixelRatio qreal as long as possible
Task-number: QTBUG-61686

+++ src/quick/items/qquickframebufferobject.cpp
@@ -260,7 +260,7 @@ public:
     bool renderPending;
     bool invalidatePending;
 
-    int devicePixelRatio;
+    qreal devicePixelRatio;
 };
 
 static inline bool isOpenGL(QSGRenderContext *rc)
